Following redirects in the TicketSharing::Client
The intention of this change is to account for ssl redirects. E.g.
'http://foo.com' => 'https://foo.com'. That is why the MAX_REDIRECTS
constant is set to 1 at the moment.

Strict protocol checking is not enforced as a redirect may change a
large portion of the url.
E.g. 'http://foo.com' => 'https://securefoo.net'

This commit also introduces a new class TicketSharing::Request. The
purpose of this class is to maintain things like the Net::HTTP
request_class and the redirect count across multiple request response
cycles.